Bug 160371: Kill the same process the second time results in ArrayIndexOutOfBoundsException and StringIndexOutOfBoundsException

+++ rse/plugins/org.eclipse.rse.services.local/src/org/eclipse/rse/services/local/processes/LocalProcessService.java
@@ -75,6 +75,12 @@ public class LocalProcessService extends AbstractProcessService implements ILoca
 		try
 		{
 			process = getProcess(monitor, PID);
+			
+			// if there is no process, simply return true
+			if (process == null) {
+				return true;
+			}
+			
 			handler.kill(process, signal);
 			return true;
 		}
